+++ src/base/b-virtual-scroll/modules/scroll-render.ts
@@ -37,6 +37,11 @@ export default class ScrollRender {
 	chunk: number = 0;
 
 	/**
+	 * Last rendered range
+	 */
+	lastRenderRange: number[] = [0, 0];
+
+	/**
 	 * Component instance
 	 */
 	readonly component: bVirtualScroll['unsafe'];
@@ -129,6 +134,7 @@ export default class ScrollRender {
 	 */
 	reInit(): void {
 		this.lastIntersectsItem = 0;
+		this.lastRenderRange = [0, 0];
 		this.chunk = 0;
 		this.items = [];
 
@@ -163,9 +169,16 @@ export default class ScrollRender {
 			renderTo = chunk * component.chunkSize,
 			renderItems = items.slice(renderFrom, renderTo);
 
-		if (!renderItems.length) {
+		if (
+			renderFrom === this.lastRenderRange[0] &&
+			renderTo === this.lastRenderRange[1] ||
+			!renderItems.length
+		) {
 			return;
 		}
+
+		this.chunk++;
+		this.lastRenderRange = [renderFrom, renderTo];
 
 		const
 			nodes = this.renderItems(renderItems);
@@ -299,7 +312,6 @@ export default class ScrollRender {
 			this.lastIntersectsItem = index;
 
 			if (currentRender - index <= renderGap) {
-				this.chunk++;
 				this.render();
 			}
 		}